% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centroids_consensus}
\alias{centroids_consensus}
\title{Centroids for the BLCA Consensus Classes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 857 rows and 9 columns.
}
\usage{
centroids_consensus
}
\description{
A data.frame containing the centroid coordinates for the six consensus classes of
muscle-invasive bladder cancer
}
\references{
\url{https://www.biorxiv.org/content/10.1101/488460v2}
}
\author{
Aurelie Kamoun
}
\keyword{datasets}
