% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centroids_tcga}
\alias{centroids_tcga}
\title{Centroids for the BLCA TCGA Classes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1264 rows and 7 columns.
}
\usage{
centroids_tcga
}
\description{
A data.frame containing the centroid coordinates for the five TCGA classes of
muscle-invasive bladder cancer
}
\references{
\url{https://doi.org/10.1016/j.cell.2017.09.007}
}
\author{
Gordon Robertson
}
\keyword{datasets}
