% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{centroids_uromol2021}
\alias{centroids_uromol2021}
\title{Centroids for the BLCA UROMOL2021 Classes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1942 rows and 7 columns.
}
\usage{
centroids_uromol2021
}
\description{
A data.frame containing the centroid coordinates for the four UROMOL2021 classes of
non-muscle-invasive bladder cancer
}
\references{
\url{https://www.nature.com/articles/s41467-021-22465-w}
}
\author{
Sia Lindskrog
}
\keyword{datasets}
