% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_blca.R
\name{classify_blca}
\alias{classify_blca}
\title{Calculate centroid correlation for samples}
\usage{
classify_blca(
  df,
  gene_id = c("entrez", "ensembl", "hgnc"),
  classifier = c("tcga", "uromol2021", "consensus"),
  tidy = FALSE
)
}
\arguments{
\item{df}{A \code{data.frame} containing RNA expression with unique genes as rows
and samples as columns. RNA-seq data must be log-transformed.}

\item{gene_id}{Character specifying the type of gene identifiers used for the
row names or first column of \code{df}.}

\item{classifier}{Classifier to be used.}

\item{tidy}{Logical. If TRUE, assumes the first column contains the gene
identifiers. Otherwise, assumes IDs are row names}
}
\value{
A \code{tibble} containing: \describe{ \item{estimate}{Pearson correlation
of a given \code{sample} to the given \code{class} centroid} \item{conf.low,
conf.high}{low and high end of 95\% confidence interval}
\item{nearest}{centroid for which the sample has the highest correlation
to} \item{statistic}{t-statistic} \item{parameter}{degrees of freedom}
\item{sep_lvl}{(Highest Correlation - 2nd Highest
Correlation)/median(Distance to highest correlation)} }
}
\description{
Calculate centroid correlation for samples
}
\details{
If using the TCGA classifier, you must provide either Entrez IDs or
HGNC symbols. The original centroids provided for the TCGA classifier only
included Entrez and HGNC, and the conversion from Entrez to Ensembl is
ambiguous. Therefore, conversion and selection of unique genes must be
performed by the user on a case-by-case basis, perhaps using the expression
value of the gene as a guide for which genes to keep.
}
\author{
Aurelie Kamoun
}
